Add files via upload
This project is about predicting whether the review written by the user is positive or negative. Here I have used Imdb dataset which has around 50000 reviews of various movies divided into positive and negatives spread randomly across the entire dataset. I have used the Logistic Regression, Random Forest and Support Vector models for training and testing the dataset. I have used 20% of the dataset for testing and 80% of it for training. Then I have compared the accuracies of these 3 models with the help of line and box plots. After training and testing the user enters the review and all the 3 models classify the review as  negative or positive and the classification which gets the majority is selected as the true classification.